% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coca.R
\name{coca}
\alias{coca}
\title{COCA}
\usage{
coca(data, rank, verbose = F, post_svd = T)
}
\arguments{
\item{data}{Data matrix or data.frame}

\item{rank}{rank of decomposition}

\item{verbose}{Should log-likelihood be printed at every step?}

\item{post_svd}{Should decomposition be post-processed with svd?}
}
\value{
A list with the following fields:

   - \code{A}: Low-rank representation of rows
   
   - \code{B}: Low-rank representation of columns
   
   - \code{fittedEsts}: Low-rank representation of full data matrix
   
Other fields are for internal use.
}
\description{
Computes low-rank COCA decomposition of data. 
Can be used with missing data.
}
\details{
Copula-based decomposition, where all data is treated as
continuous. 

For technical details:

Fang Han and Han Liu. Semiparametric principal component analysis. 
In NIPS’12: 
Proceedings of the 26th Annual Conference on Neural Information 
Processing Systems, pages 171–179, 2012.

http://papers.nips.cc/paper/4809-semiparametric-principal-component-analysis.
}
\examples{
test_data = simProblem()$data
fit = coca(test_data, rank = 2)
}
