% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{cv_rmse}
\alias{cv_rmse}
\title{Rescaled Mean Squared Error for xpca_cv}
\usage{
cv_rmse(x_true, x_est, col_index = 1)
}
\arguments{
\item{x_true}{Column of data}

\item{x_est}{Column of estimates}

\item{col_index}{Current column index}
}
\description{
Example function for \code{xpca_cv}. 
Column by column, the original data and the 
estimates **only for the data that was left out** will be
fed into this function. As such, this function must be able to ignore
missing data
}

