% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{PCA}
\usage{
pca(data, rank, tol = 0.1, rescale = T, max_iter = 500, post_svd = T)
}
\arguments{
\item{data}{matrix or data.frame}

\item{rank}{Rank of decomposition}

\item{tol}{Rank of tolerance used to assess convergence}

\item{rescale}{Should each column be scaled?}

\item{max_iter}{Maximum iterations}

\item{post_svd}{Should decomposition be post processed with svd?}
}
\value{
A list with the following fields:

   - \code{A}: Low-rank representation of rows
   
   - \code{B}: Low-rank representation of columns
   
   - \code{fittedEsts}: Low-rank representation of full data matrix
   
Other fields are for internal use.
}
\description{
Computes low-rank PCA decomposition of data. 
Preprocessing includes centering and rescaling of each column. 
Can be used with missing data.
}
\examples{
data = simProblem()$data
fit = pca(data, rank = 2)
}
