% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simProblem.R
\name{simProblem}
\alias{simProblem}
\title{Simulate Low-Rank Copula Data}
\usage{
simProblem(
  nRow = 100,
  nCol = 100,
  sigma = 0.25,
  propBinary = 0.5,
  offset = 1,
  rank = 2
)
}
\arguments{
\item{nRow}{Number of rows}

\item{nCol}{Number of columns}

\item{sigma}{Controls amount of white noise on top of low-rank structure}

\item{propBinary}{Proportion of columns that will be binary}

\item{offset}{Global mean of low-rank structure}

\item{rank}{Dimension of underlying low-rank space}
}
\value{
A list with the following fields:

  - \code{data} Random data
  - \code{meanMat} Mean of random data
}
\description{
Simulate Low-Rank Copula Data
}
