% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{visFactors}
\alias{visFactors}
\title{Visualizing Decomposition Factors}
\usage{
visFactors(decomp, usenames = T, ...)
}
\arguments{
\item{decomp}{An MXD::pca, coca or XPCA object}

\item{usenames}{Should row/column names be plotted?}

\item{...}{additional arguments passed to `plot`}
}
\description{
Visualizing Decomposition Factors
}
