% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{xpca_cv}
\alias{xpca_cv}
\title{Cross validation for MXD methods}
\usage{
xpca_cv(data, folds = 4, methods = c("pca", "xpca", "coca", "col_mean"),
  rank = 1, lossFun = cv_rmse, per.column = NA)
}
\arguments{
\item{data}{Matrix or data frame}

\item{folds}{Number of cross-validation folds}

\item{methods}{Decomposition methods used}

\item{rank}{Rank of decomposition}

\item{lossFun}{Function for computing column by column loss}
}

