% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{xpca_ise}
\alias{xpca_ise}
\title{In Sample Error for MXD methods}
\usage{
xpca_ise(data, methods = c("pca", "xpca", "coca", "col_mean"), rank = 1,
  lossFun = cv_rmse)
}
\arguments{
\item{data}{Matrix or data frame}

\item{methods}{Decomposition methods used}

\item{rank}{Rank of decomposition}

\item{lossFun}{Function for computing column by column loss}
}

