% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{xpca_iv}
\alias{xpca_iv}
\title{Validation using interval imputations}
\usage{
xpca_iv(data, rank = 1, leaveout = 0.25, pred.level = 0.9, per.column = NaN)
}
\arguments{
\item{data}{Matrix or data frame}

\item{rank}{Rank of decomposition}

\item{leaveout}{Percent of observations to remove and validate}

\item{pred.level}{Confidence of prediction interval}

\item{per.column}{If not NaN, compute only for that column}
}
\description{
Validation using interval imputations
}
