% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coca.R
\name{coca}
\alias{coca}
\title{COCA}
\usage{
coca(data, rank, verbose = F, post_svd = T)
}
\arguments{
\item{data}{Data matrix or data.frame}

\item{rank}{rank of decomposition}

\item{verbose}{Should log-likelihood be printed at every step?}

\item{post_svd}{Should decomposition be post-processed with svd?}
}

