% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{PCA}
\usage{
pca(data, rank, tol = 0.1, rescale = T, max_iter = 500, post_svd = T)
}
\arguments{
\item{data}{matrix or data.frame}

\item{rank}{Rank of decomposition}

\item{tol}{Rank of tolerance used to assess convergence}

\item{rescale}{Should each column be scaled?}

\item{max_iter}{Maximum iterations}

\item{post_svd}{Should decomposition be post processed with svd?}
}

