% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simProblem.R
\name{simProblem}
\alias{simProblem}
\title{Simulate Low-Rank Copula Data}
\usage{
simProblem(nRow = 100, nCol = 100, sigma = 0.5, propBinary = 0, rank = 3)
}
\arguments{
\item{nRow}{Number of rows}

\item{nCol}{Number of columns}

\item{sigma}{Controls amount of white noise on top of low-rank structure}

\item{propBinary}{Proportion of columns that will be binary}

\item{rank}{Dimension of underlying low-rank space}
}
\description{
Simulate Low-Rank Copula Data
}
