% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPCA.R
\name{xpca}
\alias{xpca}
\title{XPCA}
\usage{
xpca(data, rank, tol = 0.1, maxEvals = 1000, updateSigma = T,
  opt = "AN", reg_val = 0.5, post_svd = T, gridSize = nrow(data)/10)
}
\arguments{
\item{data}{Data matrix or data.frame}

\item{rank}{Rank of decomposition}

\item{maxEvals}{Maximum evaluations of loss function for L-BFGS}

\item{updateSigma}{Should sigma be updated during optimization or after?}

\item{reg_val}{Value used to regularize max intervals.}

\item{post_svd}{Should svd be used to standardize decomposition?}

\item{gridSize}{Number of points used to approximate mean function}
}

